Refactor: extract uid from currentUser
