+++ src/constants/bottomSheet.ts
@@ -1,0 +1,6 @@
+import { SnackbarOrigin } from "@mui/material";
+
+export const anchorOrigin: SnackbarOrigin = {
+  vertical: "bottom",
+  horizontal: "center",
+};
