+++ src/pages/auth/index.tsx
@@ -6,11 +6,21 @@ import Image from "next/image";
 import { useTheme } from "@mui/system";
 import { useEffect, useState } from "react";
 import { useAuth } from "src/shared/hooks/useAuth";
+import BottomSheet from "src/shared/components/bottom-sheet";
+import UserShareCode from "src/components/common/UserShareCode";
+import { useSnackbar } from "notistack";
+import { anchorOrigin } from "src/constants/bottomSheet";
 
 export default function AuthPage() {
   const theme = useTheme();
   const [isLoading, setIsLoading] = useState(false);
-  const { isAuthorized, currentUser } = useAuth();
+  const { isAuthorized, uid } = useAuth();
+  const [open, setOpen] = useState(false);
+  const { enqueueSnackbar } = useSnackbar();
+
+  const toggleDrawer = (newOpen: boolean) => () => {
+    setOpen(newOpen);
+  };
 
   const handleClick = () => {
     setIsLoading(true); //단순히 여기서만 true로 처리하고 넘어가도 될지?
@@ -19,7 +29,11 @@ export default function AuthPage() {
 
   useEffect(() => {
     if (isAuthorized) {
-      setUser(currentUser?.uid ?? "");
+      enqueueSnackbar("유저 정보를 저장 중입니다.", {
+        variant: "default",
+        anchorOrigin,
+      });
+      setUser(uid).then(() => setOpen(true));
     }
   }, [isAuthorized]);
 
@@ -42,6 +56,13 @@ export default function AuthPage() {
         <Button onClick={handleClick} isLoading={isLoading}>
           구글로 로그인하기
         </Button>
+        <BottomSheet
+          open={open}
+          onClose={toggleDrawer(false)}
+          onOpen={toggleDrawer(true)}
+        >
+          <UserShareCode />
+        </BottomSheet>
       </Container>
       <style jsx>{`
         .spacing {
